requirements.txt
pytest==7.2.0
apache-airflow-providers-docker>=3.0.0

# Instalado diretamente no Dockerfile via pip install -e /opt/framework-dataflow
